refactor(storybook): move renderSetup util
so it can be used by non-input components like buttons and dialogs
