+++ src/components/inputs/date/DateSimple.stories.js
@@ -1,6 +1,7 @@
 import ZoaDateSimple from './DateSimple.vue';
 import { ZoaInput } from '../../index.js';
-import { argTypes, renderSetup } from '../stories.js';
+import { argTypes } from '../stories.js';
+import { renderSetup } from '../../utils/stories.js';
 
 const template = `
 <zoa-input zoa-type="date-simple"
