+++ src/components/inputs/radio/Radio.stories.js
@@ -1,6 +1,7 @@
 import ZoaRadio from './Radio.vue';
 import { ZoaInput } from '../../index.js';
-import { argTypes, renderSetup } from '../stories.js';
+import { argTypes } from '../stories.js';
+import { renderSetup } from '../../utils/stories.js';
 
 const template = `
 <zoa-input zoa-type="radio"
