+++ src/components/inputs/stories.js
@@ -63,9 +63,3 @@ export const argTypes = {
     },
   },
 };
-
-export function renderSetup(args) {
-  args['rootClass'] = args.class;
-  delete args.class;
-  return args;
-}
