+++ src/components/inputs/textbox/AutocompleteTextbox.stories.js
@@ -1,6 +1,7 @@
 import ZoaAutocompleteTextbox from './AutocompleteTextbox.vue';
 import { ZoaInput } from '../../index.js';
-import { argTypes, renderSetup } from '../stories.js';
+import { argTypes } from '../stories.js';
+import { renderSetup } from '../../utils/stories.js';
 
 const template = `
 <zoa-input zoa-type="autocomplete-textbox"
