+++ src/components/inputs/textbox/Textbox.stories.js
@@ -1,6 +1,7 @@
 import ZoaTextbox from './Textbox.vue';
 import { ZoaInput } from '../../index.js';
-import { argTypes, renderSetup } from '../stories.js';
+import { argTypes } from '../stories.js';
+import { renderSetup } from '../../utils/stories.js';
 
 const template = `
 <zoa-input zoa-type="textbox"
