+++ src/components/utils/stories.js
@@ -1,0 +1,5 @@
+export function renderSetup(args) {
+  args['rootClass'] = args.class;
+  delete args.class;
+  return args;
+}
